refactor(menutitle): renamed Menutitle component to MenuTitle
